requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.11.5
chardet==3.0.4
click==8.0.4
cryptography==3.2.1
dataclasses==0.8
dbus-python==1.2.4
decorator==4.2.1
Django==3.2.25
django-filter==21.1
djangorestframework==3.15.1
file-magic==0.3.0
Flask==2.0.3
gpg==1.13.1
html5lib==0.999999999
idna==2.5
importlib-metadata==4.8.3
importlib-resources==5.4.0
isc==2.0
itsdangerous==2.0.1
Jinja2==3.0.3
libcomps==0.1.18
lxml==4.2.3
Markdown==3.3.7
MarkupSafe==2.0.1
netifaces==0.10.6
nftables==0.1
pciutils==2.3.6
perf==0.1
pexpect==4.3.1
ply==3.9
psutil==5.4.3
ptyprocess==0.5.2
pycairo==1.16.3
pycparser==2.14
pydbus==0.6.0
pygobject==3.28.3
pylve==2.1
pyOpenSSL==19.0.0
pyparsing==2.1.10
PySocks==1.6.8
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-linux-procfs==0.7.0
pytz==2024.1
pyudev==0.21.0
PyYAML==3.12
requests==2.20.0
rhnlib==2.8.6
rpm==4.14.3
selinux==2.9
sepolicy==1.1
setools==4.3.0
setroubleshoot==1.1
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.6.0
sqlparse==0.4.4
SSSDConfig==2.9.1
syspurpose==1.28.40
systemd-python==234
typing-extensions==4.1.1
urllib3==1.24.2
webencodings==0.5.1
Werkzeug==2.0.3
zipp==3.6.0
